requirements.txt
acres==0.5.0
aistudio-sdk==0.3.8
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
bce-python-sdk==0.9.52
beautifulsoup4==4.14.2
cachetools==6.2.1
camelot-py==1.0.9
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
colorlog==6.10.1
comm==0.2.3
comtypes==1.4.13
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dotenv==0.9.9
easyocr==1.7.2
et_xmlfile==2.0.0
etelemetry==0.3.1
executing==2.2.1
ezdxf==1.4.3
filelock==3.20.0
fitz==0.0.1.dev2
fonttools==4.60.1
fsspec==2025.10.0
future==1.0.0
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.42.1
google-auth-httplib2==0.2.1
google-genai==1.47.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
ImageIO==2.37.2
imagesize==1.4.1
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
# Editable install with no version control (lightglue==0.0)
-e /media/waqas/Internship/McCormickSystemsinc/dxf_approach/under use/dxf_files/LightGlue
looseversion==1.3.0
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
modelscope==1.31.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
nibabel==5.3.2
ninja==1.13.0
nipype==1.10.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
paddleocr==3.3.2
paddlex==3.3.9
pandas==2.3.3
parso==0.8.5
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.1
polars-runtime-32==1.35.1
prettytable==3.17.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
prov==2.1.1
psutil==7.1.2
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.30
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyautocad==0.2.0
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
pydot==4.0.1
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdf==5.9.0
pypdfium2==5.0.0
pytesseract==0.3.13
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==27.1.0
rdflib==7.4.0
requests==2.32.5
rsa==4.9.1
ruamel.yaml==0.18.16
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
segment-anything==1.0
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.10.16
torch==2.9.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
ultralytics==8.3.224
ultralytics-thop==2.0.18
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.14
websockets==15.0.1
